refactor(cli): move openUrlInBrowser to utils
